git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/performance@1072522 13f79535-47bb-0310-9956-ffa450edef68

+++ tools/dev/svnraisetreeconflict/main.c
@@ -192,7 +192,7 @@ raise_tree_conflict(int argc, const char **argv, apr_pool_t *pool)
   const char *wc_path, *wc_abspath;
   const char *repos_url1, *repos_url2, *path_in_repos1, *path_in_repos2;
   int operation, action, reason;
-  int peg_rev1, peg_rev2;
+  long peg_rev1, peg_rev2;
   int kind, kind1, kind2;
 
   if (argc != 13)
